refactor(contacts): unify contact creation logic and align with app conventions
