+++ core/api/src/app/accounts/index.ts
@@ -1,5 +1,6 @@
 import { AccountsRepository, WalletsRepository } from "@/services/mongoose"
 
+export * from "./create-contact"
 export * from "./account-limit"
 export * from "./create-account"
 export * from "./get-account-transactions-for-contact"
