--- core/api/src/app/contacts/index.ts
@@ -1,2 +1,0 @@
-export * from "./contact-create"
-export * from "./intraledger-contact-create"
